Split review queue runtime tests
